Avoid multiple inheritance of Track from QObject
